Extract Yaml schema from main Yaml config module
